Fixing the bugs in logging
git-svn-id: https://svn.zynga.com/svn/zcc/incubator/php-pecl-memcache-zynga/trunk@27246 93ba844f-b7fb-4add-93bc-7020b7bd55bf

+++ include/logger.h
@@ -65,7 +65,7 @@ typedef struct mc_logger : public Timer {
         log_name = ln;
     } 
 
-    void setCas(unsigned int c) {
+    void setCas(unsigned long c) {
         cas = c;
     }
